+++ lib/clipboard/java.rb
@@ -13,7 +13,8 @@ module Clipboard
     def copy(text)
       selection_string = ::Java::JavaAwtDatatransfer::StringSelection.new text
       ::Java::JavaAwt::Toolkit.default_toolkit.system_clipboard.set_contents selection_string, nil
-      paste
+
+      true
     end
 
     def paste(_ = nil)