+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.0.0 (unreleased)
 
 ### Major Changes
+* Set required Ruby to 3.0
 * `Clipboard.copy` and `.clear` now always return true instead of the pasted string
 
 ### Refactorings / Minor API Changes